requirements.txt
antlr4-python3-runtime==4.9.3
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chatterbox-tts==0.1.1
conformer==0.3.2
decorator==5.2.1
diffusers==0.29.0
distlib==0.3.9
einops==0.8.1
filelock==3.18.0
fsspec==2025.5.1
hf-xet==1.1.2
huggingface-hub==0.32.3
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.5
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.18.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pre_commit==4.2.0
protobuf==6.31.1
pycparser==2.22
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
resemble-perth==1.0.1
s3tokenizer==0.1.7
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.46.3
triton==3.2.0
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
zipp==3.22.0
